refactor(checker): split flow-type query entry points out of control_flow core
The arch size ratchet caps flow/control_flow/core.rs at 1848 lines; the
flow-query depth guard pushed it to 1888. Move the public get_flow_type
entry, its depth-guarded uncorrelated path, and the correlated
destructured-binding narrowing helpers verbatim into core/flow_query.rs
(a child module, so parent-private helpers stay accessible). core.rs drops
to ~1630 lines. No behavior change.
